fix: changed requires_grad to true for all joints
Members attached to fix joints weren't being constraind because the tensor associated with the joint had the same gradient status as the fixed (track_grad) status on the joint. Wether a joint is moved or not depends on the fixed (track_grad) property so all joints now have requires_grad as true.